Add support for Webpack 4 asset stats
In Webpack 4 all modules added using the `webpack.PrefetchPlugin` also
get a reason object explaining why the assets was added. The reason
object will have a `type: "prefetch"` property indicating that the
`PrefetchPlugin` added it, and potentially more reasons if the asset was
also required by other modules.

All modules also have a `prefetched` property indicating whether or not
it was prefetched, which we could use instead of checking `reasons`. But
that would also apply to any dynamically required modules using the
`webpackPrefetch` comment hint. Thus to keep the logic as close as
possible to how it worked in Webpack 3 this only checks for a Webpack 4
`prefetch` reason object.